+++ chevah/compat/nt_filesystem.py
@@ -16,7 +16,7 @@ from chevah.compat.exceptions import CompatError, CompatException
 from chevah.compat.helpers import _
 from chevah.compat.interfaces import ILocalFilesystem
 from chevah.compat.nt_capabilities import NTProcessCapabilities
-from chevah.compat.nt_users import NTUsers
+from chevah.compat.nt_users import NTDefaultAvatar, NTUsers
 from chevah.compat.posix_filesystem import PosixFilesystemBase
 
 
@@ -29,20 +29,6 @@ from chevah.compat.posix_filesystem import PosixFilesystemBase
 # 5 Compact Disc
 # 6 RAM Disk
 LOCAL_DRIVE = 3
-
-
-def raise_failed_to_set_owner(owner, path, message=u''):
-    '''Helper for raising the exception from a single place.'''
-    raise CompatError(1016,
-        _(u'Failed to set owner to "%s" for "%s". %s' % (
-            owner, path, message)))
-
-
-def raise_failed_to_add_group(group, path, message=u''):
-    '''Helper for raising the exception from a single place.'''
-    raise CompatError(1017,
-        _(u'Failed to add group "%s" for "%s". %s' % (
-            group, path, message)))
 
 
 class NTFilesystem(PosixFilesystemBase):
@@ -93,6 +79,20 @@ class NTFilesystem(PosixFilesystemBase):
         # Fix folder separators.
         path = path.replace('/', '\\')
         return path
+
+    @property
+    def temp_segments(self):
+        """
+        Segments to temporary folder.
+        """
+        # FIXME:930:
+        # For impersonated account we can not return the default temporary
+        # folder, which is located in default account temp folder, since
+        # impersonated account don't have access to it.
+        if not isinstance(self._avatar, NTDefaultAvatar):
+            return [u'c', u'temp']
+        else:
+            return super(NTFilesystem, self).temp_segments
 
     @classmethod
     def getEncodedPath(cls, path):
@@ -255,7 +255,7 @@ class NTFilesystem(PosixFilesystemBase):
         try:
             self._setOwner(path, owner)
         except CompatException, error:
-            raise_failed_to_set_owner(owner, path, error.message)
+            self.raiseFailedToSetOwner(owner, path, error.message)
 
     def _setOwner(self, path, owner):
         """
@@ -303,11 +303,11 @@ class NTFilesystem(PosixFilesystemBase):
                     )
             except win32net.error, error:
                 if error.winerror == 1332:
-                    raise_failed_to_set_owner(owner, path, u'No such owner.')
+                    self.raiseFailedToSetOwner(owner, path, u'No such owner.')
                 if error.winerror == 1307:
-                    raise_failed_to_set_owner(owner, path, u'Not permitted.')
+                    self.raiseFailedToSetOwner(owner, path, u'Not permitted.')
                 else:
-                    raise OSError(error.winerror, error.strerror)
+                    self.raiseFailedToSetOwner(owner, path, unicode(error))
 
     def getOwner(self, segments):
         '''See `ILocalFilesystem`.'''
@@ -333,7 +333,8 @@ class NTFilesystem(PosixFilesystemBase):
             group_sid, group_domain, group_type = (
                 win32security.LookupAccountName(None, group))
         except win32net.error:
-            raise_failed_to_add_group(group, path, u'Could not get group ID.')
+            self.raiseFailedToAddGroup(
+                group, path, u'Could not get group ID.')
 
         with self._impersonateUser():
             try:
@@ -348,7 +349,7 @@ class NTFilesystem(PosixFilesystemBase):
                 win32security.SetFileSecurity(
                     path, win32security.DACL_SECURITY_INFORMATION, security)
             except win32net.error, error:
-                raise_failed_to_add_group(
+                self.raiseFailedToAddGroup(
                     group, path, u'%s: %s' % (error.winerror, error.strerror))
 
     def removeGroup(self, segments, group):