+++ chevah/compat/tests/__init__.py
@@ -1,0 +1,23 @@
+# Copyright (c) 2011 Adi Roiban.
+# See LICENSE for details.
+"""
+Tests for the system compatibility module.
+"""
+from chevah.empirical import mk
+
+
+def setup_package():
+    """
+    Called before running all tests.
+    """
+    # Prepare the main testing filesystem.
+    mk.fs.setUpTemporaryFolder()
+
+
+def teardown_package():
+    """
+    Called after all tests were run.
+    """
+    # Remove main testing folder.
+    mk.fs.tearDownTemporaryFolder()
+    mk.fs.checkCleanTemporaryFolders()